Search for commands from shortest to longest.
The "PathName" of a service is actually the full path to the service's
executable and all of its arguments (if any).  In order to figure out
what may need to be quoted, the actual executable must be found in the
string and separated from its arguments.  There are two ways to do this:
either iterate from the longest "path" to the shortest, or from the
shortest to the longest.  Longest-to-shortest (I think) is more secure,
at the expense of taking longer to find the actual command.  Shortest-to-
longest is less secure, since an attacker *could* have already dropped
"C:\Program.exe" onto the system (which would then be shorter than
"C:\Program Files\MyCoolService\service.exe").  But if you're running this
cmdlet, you should probably also be verifying the output of it and not
blindly piping it to Repair-ServicePath.  Doing shortest-to-longest
checking takes the cmdlet procesing time on a "normal" computer from 15-18
seconds down to 2.5 seconds.